Spacing, un-authoring for ambiguity
Added some additional comments and also removing the authorship stuff as I assume it doesn't fully matter under GNU. 
Allowing for multiple urls in settings file. Just add the beginning of the Url for instance:
Penn state is https://webaccess.psu.edu/cgi-bin/logout, however if your logout url is different fill it in and the rest will be automated.